pyproject.toml
[tool.poetry]
name = "copilot-more"
version = "0.1.0"
description = ""
authors = ["Jijun Leng <jijunleng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
aiohttp = "^3.11.8"
mitmproxy = "^11.0.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
